Move chunk update event handling into map.data.classic.
